services/vision-analyzer-service/pyproject.toml
[project]
name = "vision-analyzer-service"
version = "0.1.0"
description = "Analyze frames for IP infringement using Gemini 2.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "uvicorn[standard]>=0.38.0",
    "pydantic-settings>=2.11.0",
    "google-genai>=1.46.0",
    "pillow>=12.0.0",
    "google-cloud-storage>=3.5.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-pubsub>=2.33.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-secret-manager>=2.25.0",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
